% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_filter_rank.R
\name{filter_rank}
\alias{filter_rank}
\title{Function to download species from fishtreeoflife}
\usage{
filter_rank(order)
}
\arguments{
\item{order}{Character with names of species}
}
\value{
phylogeny
}
\description{
Function to download species from fishtreeoflife
}
