% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{fish_SAmerica}
\alias{fish_SAmerica}
\title{Occurrence of fish species in South America}
\format{
A data frame with 37 rows and 679 variables:
}
\usage{
fish_SAmerica
}
\description{
A dataset containing the occurrence of  fish species distributed all basins of
South America continent
}
\keyword{datasets}
