% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishMakeNodeMaker.R
\name{FishNodeNameMaker}
\alias{FishNodeNameMaker}
\title{Title}
\usage{
FishNodeNameMaker(phylo, data)
}
\arguments{
\item{phylo}{A phylo object}

\item{data}{A three column data frame containing the name of species and their family and order.
can be obtained using 'FishTaxaMaker' function}
}
\value{
A phylo objetct with node names
}
\description{
Title
}
