% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishPhyloMaker.R
\name{FishPhyloMaker}
\alias{FishPhyloMaker}
\title{Editing fish phylogeny according to a local pool of species}
\usage{
FishPhyloMaker(data, return.insertions = FALSE)
}
\arguments{
\item{data}{A data frame with three columns containing the name of species (s), the Family (f) and the Order (o). This data frame can be generated
with tab_function function}

\item{return.insertions}{Logical, if TRUE the output is a list of length two containing the phylogeny
and a dataframe with a column indicating at which level each species was inserted}
}
\value{

}
\description{
Editing fish phylogeny according to a local pool of species
}
\examples{
\dontrun{
    data(neotropical_comm)
    data_comm <- neotropical_comm[, -c(1, 2)]
    taxon_data <- tab_function(data_comm) # formating data
    Loricariidae # informing Family of Curculionichthys insperatus
    Siluriformes # informing Order of Curculionichthys insperatus
    res_test <- FishPhyloMaker(data = data, return.insertions = TRUE)
    res_test$Phylogeny # newick object
    res_test$insertions # dataframe containing the level in which each species was inserted
}


}
