% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{neotropical_comm}
\alias{neotropical_comm}
\title{Abundance of stream fish species in Parana and Paraguay streams}
\format{
A data frame with 20 rows and 61 variables:
}
\usage{
neotropical_comm
}
\description{
A dataset containing the abundance of stream fish species distributed in streams of
Parana and Paraguay river Basins
}
\references{
\url{https://doi.org/10.1590/1982-0224-2020-0126}
}
\keyword{datasets}
