% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_deficit_2.R
\name{PD_deficit_2}
\alias{PD_deficit_2}
\title{Calculate PD deficit at species and insertion levels}
\usage{
PD_deficit_2(phylo, data)
}
\arguments{
\item{phylo}{a phylo object}

\item{data}{data frame with species and insertion levels. This is the output of FishPhyloMaker function}
}
\value{
a list with two tibbles. One with results of deficit for all species and another with summary results for each category of insertion

a list containing two data frames.

\if{html}{\out{<div class="sourceCode">}}\preformatted{- a long data frame (pd_deficit_long) containing all values of deficit and total phylogenetic diversity
    accounted by each terminal branch of the tree.

- a summary data frame with values of total PD in each category of insertion (including for those species present in tree)
    and a proportion of each category for the total terminal PD.
}\if{html}{\out{</div>}}
}
\description{
Calculate PD deficit at species and insertion levels
}
