% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_deficit.R
\name{PD_defict}
\alias{PD_defict}
\title{Title Calculate the amount of phylogenetic deficit in assemblages}
\usage{
PD_defict(phylo, data, level = "Congeneric_insertion")
}
\arguments{
\item{phylo}{Phylogenetic tree in newick format, can be an object from \code{\link{FishPhyloMaker}} function}

\item{data}{A data frame containing the classification  informing the level of insertions. This can be obtained
from \code{\link{FishPhyloMaker}} function}

\item{level}{Character indicating which level must be considered in the calculation of PD deficit.
default is "Congeneric_insertion"}
}
\value{
A scalar containing the value of PD deficit for the level chosen
}
\description{
Title Calculate the amount of phylogenetic deficit in assemblages
}
\seealso{
\code{\link{FishPhyloMaker}} for phylogeny and data frame containing the classification of insertions
}
