% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishPhyloMaker.R
\name{phyloMatch}
\alias{phyloMatch}
\title{Making fish phylogeny for a local pool of species}
\usage{
phyloMatch(data)
}
\arguments{
\item{data}{A character vector containing the names of species or a dataframe generated from \code{tab_function}}
}
\value{
A newick object describing the phylogenetic relationships among the species provided in \code{data}
}
\description{
Making fish phylogeny for a local pool of species
}
\examples{

}
