% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_user_opt_printCat.R
\name{print_cat}
\alias{print_cat}
\title{Internal function - auxiliar to interactive procedure}
\usage{
print_cat(print_cat, spp, family)
}
\arguments{
\item{print_cat}{Character}

\item{spp}{Character}

\item{family}{Character}
}
\value{
interactive action in console
}
\description{
Internal function - auxiliar to interactive procedure
}
