% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_user_opt_printCat2.R
\name{print_cat2}
\alias{print_cat2}
\title{Internal function}
\usage{
print_cat2(spp)
}
\arguments{
\item{spp}{Character}
}
\value{
interactive action in console
}
\description{
Internal function
}
