% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_user_opt_printCatFamily.R
\name{print_cat_family}
\alias{print_cat_family}
\title{Internal function to help in interactive process}
\usage{
print_cat_family(print_cat, spp, order)
}
\arguments{
\item{print_cat}{Character}

\item{spp}{Character}

\item{order}{Character}
}
\value{
interactive action in consele
}
\description{
Internal function to help in interactive process
}
