% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{spp_afrotropic}
\alias{spp_afrotropic}
\title{List of fish species with occurrence in Afrotropical ecoregion}
\format{
A character vector with 767 species names:
}
\usage{
spp_afrotropic
}
\description{
A list of species that occur in basins of
Afrotropical ecoregion
}
\references{
\url{https://www.nature.com/articles/sdata2017141}
}
\keyword{datasets}
