% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{taxon_data_PhyloMaker}
\alias{taxon_data_PhyloMaker}
\title{Data frame with species names needed to assemble the phylogenetic tree}
\format{
A data frame with taxonomic classification (species, family and order) of 45 species
}
\usage{
taxon_data_PhyloMaker
}
\description{
A data frame that can be directly used in FishPhyloMaker to obtain a phylogenetic tree
}
\references{
Species that make up the dataset in the paper published in Neotropical Ichthyology  \doi{10.1590/1982-0224-2020-0126}
}
\keyword{datasets}
