% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_treedata_modif.R
\name{treedata_modif}
\alias{treedata_modif}
\title{Internal function to show the number of species from species pool that lacks in phylogeny}
\usage{
treedata_modif(phy, data, sort = FALSE, warnings = TRUE)
}
\arguments{
\item{phy}{Phylogenetic hypothesis in newick format}

\item{data}{Data frame with species to be added in tree}

\item{sort}{Sorting species in alphabetic order}

\item{warnings}{Logical}
}
\value{
character vector
}
\description{
Internal function to show the number of species from species pool that lacks in phylogeny
}
