% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_function.R
\name{tab_function}
\alias{tab_function}
\title{Generate a list of species
Auxiliar function to ensemble the data frame of species pool}
\usage{
tab_function(data)
}
\arguments{
\item{data}{A character vector with species names or a community matrix with species names in columns}
}
\value{

}
\description{
Generate a list of species
Auxiliar function to ensemble the data frame of species pool
}
\examples{

}
